Use external crypto crates
