--- rust_blowfish/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "rust_blowfish"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-libc = "0.2"
-blowfish = "0.8"
-pbkdf2 = "0.12"
-hmac = "0.12"
-sha2 = "0.10"
-hex = "0.4"
-
-[lib]
-name = "rust_blowfish"
-crate-type = ["staticlib", "rlib"]
