--- rust_crypt/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "rust_crypt"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "rust_crypt"
-crate-type = ["staticlib", "rlib"]
-
-[dependencies]
-blowfish = "0.8"
-ring = "0.17"
-libc = "0.2"
